+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to `laravel-site-settings` will be documented in this file
 
-## 1.0.0 - 201X-XX-XX
+## 1.0.0 - 2019-04-04
 
 - initial release
+
+## 1.0.1 - 2019-24-04
+
+- Add command to create options
+
+## 1.0.2 - 2019-30-04
+
+- Change model name from LaravelSiteSettings to Setting
+- Improve Readme.md
+- Fix some minor issues